Bluetooth: Audio: Move endpoint.c into bap.c
The bap.c file is the current file for the
unicast client implementation, and the remaining
part of endpoint.c is for the unicast client
only, so the two files are merged.

Code is not updated with naming in this commit, as to avoid
any necessary code changes for review.

The removes the "endpoint" implementation completely.
