+++ subsys/bluetooth/host/audio/CMakeLists.txt
@@ -18,7 +18,6 @@ zephyr_library_sources_ifdef(CONFIG_BT_AUDIO_STREAM stream.c)
 zephyr_library_sources_ifdef(CONFIG_BT_AUDIO_BROADCAST_SOURCE broadcast_source.c)
 zephyr_library_sources_ifdef(CONFIG_BT_AUDIO_BROADCAST_SINK broadcast_sink.c)
 zephyr_library_sources_ifdef(CONFIG_BT_AUDIO_UNICAST_SERVER unicast_server.c)
-zephyr_library_sources_ifdef(CONFIG_BT_AUDIO_ENDPOINT endpoint.c)
 zephyr_library_sources_ifdef(CONFIG_BT_AUDIO_CAPABILITIES capabilities.c)
 zephyr_library_sources_ifdef(CONFIG_BT_ASCS ascs.c)
 zephyr_library_sources_ifdef(CONFIG_BT_BAP bap.c)
