+++ subsys/bluetooth/host/audio/Kconfig
@@ -80,13 +80,6 @@ config BT_AUDIO_DEBUG_STREAM
 	  Use this option to enable Bluetooth Audio Stream debug logs for the
 	  Bluetooth Audio functionality.
 
-config BT_AUDIO_DEBUG_ENDPOINT
-	bool "Bluetooth Audio Endpoint debug"
-	depends on BT_AUDIO_ENDPOINT
-	help
-	  Use this option to enable Bluetooth Audio Endpoint debug logs for the
-	  Bluetooth Audio functionality.
-
 config BT_AUDIO_DEBUG_CAPABILITIES
 	bool "Bluetooth Audio Capabilities debug"
 	depends on BT_AUDIO_CAPABILITIES
