+++ subsys/bluetooth/host/audio/Kconfig.baps
@@ -10,11 +10,6 @@ config BT_AUDIO_STREAM
 	bool
 	default y if BT_ASCS || BT_BAP || BT_AUDIO_BROADCAST_SINK || BT_AUDIO_BROADCAST_SOURCE
 
-config BT_AUDIO_ENDPOINT
-	# Virtual/hidden option
-	bool
-	default y if BT_BAP
-
 config BT_AUDIO_CAPABILITIES
 	# Virtual/hidden option
 	bool
