+++ tests/bluetooth/audio_shell/prj.conf
@@ -102,7 +102,6 @@ CONFIG_BT_DEBUG_TBS=y
 CONFIG_BT_DEBUG_CCP=y
 CONFIG_BT_DEBUG_PACS=y
 CONFIG_BT_AUDIO_DEBUG_STREAM=y
-CONFIG_BT_AUDIO_DEBUG_ENDPOINT=y
 CONFIG_BT_AUDIO_DEBUG_CAPABILITIES=y
 CONFIG_BT_AUDIO_DEBUG_BROADCAST_SOURCE=y
 CONFIG_BT_AUDIO_DEBUG_BROADCAST_SINK=y
