+++ tests/bluetooth/bsim_bt/bsim_test_audio/prj.conf
@@ -111,7 +111,6 @@ CONFIG_BT_DEBUG_BASS_CLIENT=n
 CONFIG_BT_AUDIO_DEBUG_UNICAST_SERVER=y
 CONFIG_BT_AUDIO_DEBUG_STREAM=y
 CONFIG_BT_AUDIO_DEBUG_CAPABILITIES=y
-CONFIG_BT_AUDIO_DEBUG_ENDPOINT=y
 CONFIG_BT_AUDIO_DEBUG_BROADCAST_SOURCE=y
 CONFIG_BT_AUDIO_DEBUG_BROADCAST_SINK=y
 CONFIG_BT_DEBUG_BAP=y
